refactor: WithmoButtonを作成し、そこで連続タップできないように変更
